# docs/swagger.yaml
basePath: /api/v1
definitions:
  apperrors.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  apperrors.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  model.TripStop:
    properties:
      arrival_time:
        type: string
      departure_time:
        type: string
      id:
        type: integer
      price_to_next:
        type: integer
      stop_id:
        type: integer
      stop_order:
        type: integer
      trip_id:
        type: integer
    type: object
  model.TripStopUpdate:
    properties:
      arrival_time:
        type: string
      departure_time:
        type: string
      price_to_next:
        type: integer
      stop_id:
        type: integer
      stop_order:
        type: integer
      trip_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: luchitstimofei@yandex.ru
    name: Luchits Timofei
    url: http://example.com
  description: This is a server for order private bus.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ordering Bus API
  version: "1.0"
paths:
...
  /admin/trip_stops:
    post:
      consumes:
      - application/json
      description: Создает новую запись остановки (требуются права администратора)
      parameters:
      - description: Данные остановки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.TripStop'
      produces:
      - application/json
      responses:
        "201":
          description: Остановка успешно создана
          schema:
            $ref: '#/definitions/apperrors.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - Bearer: []
      summary: Создать новую остановки (только админ)
      tags:
      - admin/trip_stops
  /admin/trip_stops/{id}:
    delete:
      description: Удаляет запись об остановке по ID (требуются права администратора)
      parameters:
      - description: ID остановки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Остановка успешно удалена
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "404":
          description: Остановка не найдена
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - Bearer: []
      summary: Удалить остановки (только админ)
      tags:
      - admin/trip_stops
    put:
      consumes:
      - application/json
      description: Обновляет информацию об остановке по ID (требуются права администратора)
      parameters:
      - description: ID остановки
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные остановки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.TripStopUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Данные остановки обновлены
          schema:
            $ref: '#/definitions/apperrors.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "404":
          description: Водитель не найден
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - Bearer: []
      summary: Обновить данные об остановки (только админ)
      tags:
      - admin/trip_stops
...
  /trip_stops:
    get:
      description: Возвращает все остановки в системе
      produces:
      - application/json
      responses:
        "200":
          description: Модель остановок
          schema:
            $ref: '#/definitions/model.TripStop'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Получить все остановки
      tags:
      - trip_stops
  /trip_stops/{id}:
    get:
      description: Возвращает остановку по ID
      parameters:
      - description: ID остановки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Модель пути
          schema:
            $ref: '#/definitions/model.TripStop'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Получить остановку
      tags:
      - trip_stops
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"